6 outta 6 features

See merge request cse1105/2018-2019/oopp-group-83/template!64

+++ src/main/java/calculator/CarCalculator.java
@@ -6,9 +6,16 @@ import org.json.JSONObject;
 import java.net.MalformedURLException;
 import java.net.URL;
 
+
 public class CarCalculator {
 
-    public static int car(int distance) {
+    /**
+     * The calculator for getting the amount of kg CO2 produced by a car.
+     * @param distance The distance
+     * @return Amount of kg CO2 produced by a car
+     */
+
+    public static Double car(int distance) {
 
         //Setup https client
         String host = "http://impact.brighterplanet.com/";
@@ -40,8 +47,7 @@ public class CarCalculator {
         //Calculate amount of kg CO2 saved.
         Double co2Car = third.getDouble("value");
         co2Car = Math.round(co2Car * 100.00) / 100.00;
-        return co2Car.intValue();
-
+        return co2Car;
     }
 
 }
